stream refactoring fixed cascade peptidehasmodification annotation

+++ colims-repository/src/test/java/com/compomics/colims/repository/ExperimentRepositoryTest.java
@@ -29,19 +29,19 @@ public class ExperimentRepositoryTest {
         Experiment experiment = experimentRepository.findById(1L);
         Long count = experimentRepository.countByProjectIdAndTitle(1L, experiment);
 
-        Assert.assertTrue(count.longValue() == 0);
+        Assert.assertTrue(count == 0);
 
         //try to find an non existing experiment within an existing project
         experiment = new Experiment();
         experiment.setTitle("Unknown experiment 1 title");
         count = experimentRepository.countByProjectIdAndTitle(1L, experiment);
 
-        Assert.assertTrue(count.longValue() == 0);
+        Assert.assertTrue(count == 0);
 
         //try to find an non existing experiment within a non existing project
         count = experimentRepository.countByProjectIdAndTitle(44L, experiment);
 
-        Assert.assertTrue(count.longValue() == 0);
+        Assert.assertTrue(count == 0);
     }
 
 }
